Change X-MEDIA-SEQUENCE type from int to long Failing playlist: https://kitagaruda.blob.core.windows.net/m3u8-files/26158/240p/playlist.m3u8 contains value: #EXT-X-MEDIA-SEQUENCE:15385550390000

+++ src/main/java/io/lindstrom/m3u8/model/MediaPlaylist.java
@@ -13,7 +13,7 @@ public interface MediaPlaylist extends Playlist {
     int targetDuration();
 
     @Value.Default
-    default int mediaSequence() {
+    default long mediaSequence() {
         return 0;
     }
